src/backend/requirements.txt
flask
flask-cors
openai
openpyxl
python-docx
python-dotenv
requests
beautifulsoup4
